DZNPY-0048: Relocating versioning aspects
