+++ src/dznpy/adv_shell/__init__.py
@@ -28,7 +28,7 @@ Example configurations:
 from typing import Optional, List
 
 # dznpy modules
-from ..dznpy_version import VERSION
+from ..versioning import DZNPY_VERSION
 from .. import cpp_gen
 from ..ast_view import find_fqn
 from ..cpp_gen import AccessSpecifier, Comment
@@ -237,7 +237,7 @@ class Builder:
 
         cpp.namespace.contents = TB(str(cpp.struct))
 
-        footer = Comment(f'Generated by: dznpy/adv_shell v{VERSION}')
+        footer = Comment(f'Generated by: dznpy/adv_shell v{DZNPY_VERSION}')
 
         return GeneratedContent(filename=f'{cpp.target_file_basename}.hh',
                                 contents=str(TextBlock([header, cpp.namespace, footer])))
@@ -274,7 +274,7 @@ class Builder:
                                      chunk(cpp.provides_port_helpers.private_def),
                                      ])
 
-        footer = Comment(f'Generated by: dznpy/adv_shell v{VERSION}')
+        footer = Comment(f'Generated by: dznpy/adv_shell v{DZNPY_VERSION}')
 
         return GeneratedContent(filename=f'{cpp.target_file_basename}.cc',
                                 contents=str(TextBlock([header, cpp.namespace, footer])))
