--- src/dznpy/dznpy_version.py
@@ -1,10 +1,0 @@
-"""
-dznpy - centralized version and copyright constants
-"""
-
-COPYRIGHT = '''\
-Copyright (c) 2023-2025 Michael van de Ven <michael@ftr-ict.com>
-This is free software, released under the MIT License. Refer to dznpy/LICENSE.
-'''
-
-VERSION = '1.3.DEV'
