+++ src/dznpy/support_files/__init__.py
@@ -11,7 +11,7 @@ from typing import Optional, Tuple
 
 # dznpy modules
 from ..cpp_gen import fqn_t, Namespace, Comment
-from ..dznpy_version import VERSION, COPYRIGHT
+from ..versioning import DZNPY_VERSION, DZNPY_COPYRIGHT
 from ..misc_utils import assert_t_optional, assert_t
 from ..scoping import NamespaceIds, ns_ids_t
 from ..text_gen import BLANK_LINE, chunk, DO_NOT_MODIFY, TextBlock, TB
@@ -34,7 +34,7 @@ def distillate_ns(namespace_prefix: Optional[NamespaceIds]) -> Tuple[NamespaceId
 
 def footer() -> str:
     """Generate the generic footer for support files."""
-    return str(Comment(f'Generated by: dznpy/support_files v{VERSION}'))
+    return str(Comment(f'Generated by: dznpy/support_files v{DZNPY_VERSION}'))
 
 
 @dataclass(frozen=True)
@@ -61,7 +61,7 @@ def generate_cpp_code(cfg: SupportFileCfg) -> str:
 
     extended_header = Comment([chunk(cfg.header),
                                chunk(DO_NOT_MODIFY),
-                               COPYRIGHT
+                               DZNPY_COPYRIGHT
                                ])
 
     namespace_with_body = Namespace(full_ns, contents=TB([BLANK_LINE,
