+++ test/unit_tests/support_files/test_sf_meta_helpers.py
@@ -16,7 +16,7 @@ from dznpy.support_files import meta_helpers as sut
 
 # Test data
 from common.testdata import *
-from dznpy.dznpy_version import VERSION
+from dznpy.versioning import DZNPY_VERSION
 
 
 def template_hh(ns_prefix: str) -> str:
@@ -69,7 +69,7 @@ DZN_PORT CreatePort(const std::string& provideName, const std::string& requireNa
 }
 
 } // namespace """ f'{ns_prefix}' """Dzn
-// Generated by: dznpy/support_files v"""f'{VERSION}'"""
+// Generated by: dznpy/support_files v"""f'{DZNPY_VERSION}'"""
 """
