+++ test/unit_tests/support_files/test_sf_misc_utils.py
@@ -16,7 +16,7 @@ from dznpy.support_files import misc_utils as sut
 
 # Test data
 from common.testdata import *
-from dznpy.dznpy_version import VERSION
+from dznpy.versioning import DZNPY_VERSION
 
 
 def template_hh(ns_prefix: str) -> str:
@@ -76,7 +76,7 @@ template <typename STR_TYPE>
 }
 
 } // namespace """ f'{ns_prefix}' """Dzn
-// Generated by: dznpy/support_files v"""f'{VERSION}'"""
+// Generated by: dznpy/support_files v"""f'{DZNPY_VERSION}'"""
 """
