+++ test/unit_tests/support_files/test_sf_multi_client_selector.py
@@ -16,7 +16,7 @@ from dznpy.support_files import multi_client_selector as sut
 
 # Test data
 from common.testdata import *
-from dznpy.dznpy_version import VERSION
+from dznpy.versioning import DZNPY_VERSION
 
 
 def template_hh(cpp_ns_prefix: str, file_ns_prefix: str) -> str:
@@ -184,7 +184,7 @@ private:
 };
 
 } // namespace """ f'{cpp_ns_prefix}' """Dzn
-// Generated by: dznpy/support_files v"""f'{VERSION}'"""
+// Generated by: dznpy/support_files v"""f'{DZNPY_VERSION}'"""
 """
