+++ test/unit_tests/support_files/test_sf_strict_port.py
@@ -16,7 +16,7 @@ from dznpy.support_files import strict_port as sut
 
 # Test data
 from common.testdata import *
-from dznpy.dznpy_version import VERSION
+from dznpy.versioning import DZNPY_VERSION
 
 
 def template_hh(ns_prefix: str) -> str:
@@ -87,7 +87,7 @@ void ConnectPorts(Mts<P> provided, Mts<P> required)
 }
 
 } // namespace """ f'{ns_prefix}' """Dzn
-// Generated by: dznpy/support_files v"""f'{VERSION}'"""
+// Generated by: dznpy/support_files v"""f'{DZNPY_VERSION}'"""
 """
